fix display name
